ECE 157C/HW1/requirements.txt
langgraph>=0.2.60
langchain-openai>=0.2.14
langchain-core>=0.3.20
openai>=1.55.0
pandas>=2.2.3
python-dotenv>=1.0.1
